+++ packages/flutter_fixtures_dio/lib/flutter_fixtures_dio.dart
@@ -2,3 +2,4 @@ library flutter_fixtures_dio;
 
 export 'src/dio_interceptor.dart';
 export 'src/recorder_interceptor.dart';
+export 'src/response_origin.dart';